Labeling & moderation updates to SDK (#1366)
* First pass on label definitions and reference-doc generation for them

* Tune labels

* Drop the 'csam' label in favor of using !hide

CSAM is an extremely delicate issue. Moderators must be careful to
remove it from the system without calling direct attention to it.
Using the generic !hide achieves the same effect while obscuring
the reason for the removal. Server and AppView bans will then be
employed (along with any other needed mechanisms) to strip the
content from the network.

There is also a 'false accusation' concern with using the csam
label. Fingerprinting techniques and AI are used to scan for csam.
These are capable of producing false positives which will then be
reverted after human review. The reputational damage caused by a
false positive could be severe, so we want to be careful about any
system which might publicize a false positive.

* Add label definition code generation

* Implement all type signatures for moderation

* Complete a first pass on the labeling sdk

* Add post-moderation test suite

* Add post moderation behavior documentation

* Add self-post behaviors

* Improve post moderation doc

* Tune up the post moderation doc

* Tune up the post moderation doc

* Tune up the post moderation doc

* Simplify the post moderation behavior descriptions

* More behavior description tuneup

* Add profile moderation final behaviors and tests

* Improve generated post-moderation-behaviors doc

* Add profile moderation behaviors doc

* Test muted-by-list specifically

* Fixes to label descriptions

* Fix to muted-by-list behavior

* Dont blur account in moderateProfile() when muting

* Tune label copy

* Apply post embed blurring when account is marked with blurmedia label

* Fix output signature

* Fixes to blocking behavior

* Rename LabelDefinitionPreference to LabelPreference

* Update docs

* Fix test

* Fix: self-harm should blur media only

* Fixes to tests

* Improve label copy

* Remove all labels that do not have a specific policy

Communicating moderation policies with active users is important.
This PR originally included labels which were proposed but did not
yet have policies. While we didn't intend to use them until a
policy was established, I decided it's better to hold off putting
them in the code until we're sure about them. They can be found in
backup files prefixed with "proposed-".